sheep: refactor functions shared by store_driver
This change is a preparation patch for add store_driver.
move the definition of functions that are shared by store_driver


Conflicts:
	sheep/store/plain_store.c
