+++ sheep/sheep_priv.h
@@ -389,6 +389,10 @@ int leave_cluster(void);
 
 void queue_cluster_request(struct request *req);
 
+int prepare_iocb(uint64_t oid, const struct siocb *iocb, bool create);
+int err_to_sderr(const char *path, uint64_t oid, int err);
+int discard(int fd, uint64_t start, uint32_t end);
+
 int update_epoch_log(uint32_t epoch, struct sd_node *nodes, size_t nr_nodes);
 int inc_and_log_epoch(void);
